bridge: Issue all entity updates in parallel
Calls to updateEntity may take time, especially if that entity has an
ongoing call so that there is mutex contention.

Issue all the calls in parallel and await them at the en up
updateEntities. Ordering of state updates to a particular entity should
be enforced by the mutex which queues waiters internally.