Update README.md
Hi @t0bst4r,

I added my own YT-Video (hope you like it 🎉) as a resource down in the README.md and added a section at the bottom for a better overview of additional Resources, so that others may add their Videos in the future. Do you like it?

I also corrected very little grammar in the README, but really not that much 😉. Love the work and passion you put into your Project & keep on rocking 🚀

Best Regards,
Simon