speaker: Specify min/max values
When min and max values aren't specified, Google Home sends moveToLevel
commands in the range of 0-100 when it should be 0-254, making most of
the volume range unavailable.

Setting min/max for at 0-100 range makes both app and voice controls
work as intended, but the volume displayed in the Google Home app
updates incorrectly as if it thinks it needs to convert from 0-254 to
0-100.

Still an improvement though.