fix(basic-information-server): Use device vendor
We previously kept the t0bst4r vendor on all bridged devices to support
a workaround present in Alexa Media Player that relied on the vendor to
detect if a device was bridged from home assistant to avoid loops.

As of version 5.4.0, Alexa Media Player understands the relationship
between our bridge and our bridged devices, and only needs to recognize
the vendor of the bridge itself.

Report the correct vendor for bridged devices as the workaround is no
longer relevant.